backend/requirements.txt
# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.25.0

# PDF processing
PyPDF2==3.0.1
pypdf==4.0.1

# Text processing and NLP
sentence-transformers==2.2.2
transformers==4.36.2
tokenizers==0.15.1
torch>=2.0.0,<3.0.0

# Vector database
faiss-cpu==1.7.4
numpy>=1.24.0,<2.0.0

# Data handling
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# HTTP requests for PDF downloads
requests>=2.31.0,<3.0.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# Logging and utilities
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# Evaluation metrics
nltk>=3.8.0,<4.0.0
rouge-score>=0.1.2,<1.0.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0 